Update repository with the latest changes.

+++ Headers/LinearAlgebra.h
@@ -4,13 +4,13 @@
 #include <vector>
 
 namespace LinearAlgebra {
-	std::vector<std::vector<double>> build(int const& rows, int const& cols);
+	std::vector<std::vector<double>> build(unsigned const& rows, unsigned int const& cols);
 
-	std::vector<std::vector<double>> identity(int const& rows, int const& cols);
+	std::vector<std::vector<double>> identity(unsigned int const& rows, unsigned int const& cols);
 
-	std::vector<std::vector<double>> row(double const& row, std::vector<std::vector<double>> const& a);
+	std::vector<std::vector<double>> row(unsigned int const& row, std::vector<std::vector<double>> const& a);
 
-	std::vector<std::vector<double>> column(double const& column, std::vector<std::vector<double>> const& a);
+	std::vector<std::vector<double>> column(unsigned int const& column, std::vector<std::vector<double>> const& a);
 
 	std::vector<std::vector<double>> scalar(double const& c, std::vector<std::vector<double>> a);
